[REFACTOR] Move extractor functionalities to another file
