+++ backend/src/services/invoiceMapperService.js
@@ -1,8 +1,11 @@
 'use strict';
 const FieldParser = require('./invoiceMapperService/FieldParserService');
+const EntityExtractor = require('./invoiceMapperService/entityExtractorService');
+
 class AzureInvoiceMapper {
   constructor() {
     this.fieldParser = new FieldParser();
+    this.EntityExtractor = new EntityExtractor(this.fieldParser);
   }
   /**
    * Maps Azure OCR result to Invoice model format
@@ -51,12 +54,12 @@ class AzureInvoiceMapper {
     const paymentTerms = this.fieldParser.getFieldContent(fields.PaymentTerm);
 
     // Extract line items
-    const itemsData = this.extractLineItems(fields.Items);
+    const itemsData = this.EntityExtractor.extractLineItems(fields.Items);
     
     // Extract customer data into a separate object
-    const customerData = this.extractCustomerData(fields);
+    const customerData = this.EntityExtractor.extractCustomerData(fields);
 
-    const vendorData = this.extractVendorData(fields);
+    const vendorData = this.EntityExtractor.extractVendorData(fields);
 
     // Build invoice data object matching our model requirements
     const invoiceData = {
@@ -83,82 +86,6 @@ class AzureInvoiceMapper {
       vendorData,
       itemsData
     };
-  }
-
-  /**
-   * Extract customer data from OCR fields into a structured object
-   * @param {Object} fields - OCR fields
-   * @returns {Object} Structured customer data
-   */
-  extractCustomerData(fields) {
-    const addressData = this.fieldParser.getFieldContent(fields.CustomerAddress || fields.BillingAddress || fields.ShippingAddress);
-    return {
-      name: this.fieldParser.getFieldContent(fields.CustomerName) || this.fieldParser.getFieldContent(fields.BillingAddressRecipient),
-      address: addressData, 
-      recipient_name: this.fieldParser.getFieldContent(fields.CustomerAddressRecipient) ||
-        this.fieldParser.getFieldContent(fields.CustomerName),
-      tax_id: this.fieldParser.getFieldContent(fields.CustomerTaxId) ||
-        this.fieldParser.getFieldContent(fields.VatNumber) ||
-        this.fieldParser.getFieldContent(fields.TaxId)
-    };
-  }
-  /**
- * Extract vendor data from OCR fields into a structured object
- * @param {Object} fields - OCR fields
- * @returns {Object} Structured vendor data
- */
-  extractVendorData(fields) {
-    const addressData = this.fieldParser.getFieldContent(fields.VendorAddress);
-
-    return {
-      name: this.fieldParser.getFieldContent(fields.VendorName),
-      address: addressData,
-      recipient_name: this.fieldParser.getFieldContent(fields.VendorAddressRecipient) ||
-        this.fieldParser.getFieldContent(fields.VendorName),
-      tax_id: this.fieldParser.getFieldContent(fields.VendorTaxId) ||
-        this.fieldParser.getFieldContent(fields.VendorVatNumber) ||
-        this.fieldParser.getFieldContent(fields.SupplierTaxId)
-    };
-  }
-
-
-  /**
-   * Extract line items from OCR result with improved handling
-   * @param {Object} itemsField - Items field from OCR
-   * @returns {Array} Extracted line items
-   */
-  extractLineItems(itemsField) {
-    if (!itemsField) {
-      return [];
-    }
-
-    // Handle array format
-    if (itemsField.values) {
-      return itemsField.values.map(item => {
-        const fields = item.properties || {};
-        return {
-          description: this.fieldParser.getFieldContent(fields.Description) || this.fieldParser.getFieldContent(fields.ProductCode),
-          quantity: this.fieldParser.parseNumeric(fields.Quantity),
-          unit: this.fieldParser.getFieldContent(fields.Unit),
-          unitPrice: this.fieldParser.parseCurrency(fields.UnitPrice).amount,
-          amount: this.fieldParser.parseCurrency(fields.Amount).amount,
-        };
-      });
-    }
-
-    // Sometimes the items field itself has content but not in valueArray format
-    const content = this.fieldParser.getFieldContent(itemsField);
-    if (content) {
-      return [{
-        description: content,
-        quantity: null,
-        unit: null,
-        unitPrice: null,
-        amount: null
-      }];
-    }
-
-    return [];
   }
 }
