+++ backend/tests/services/InvoiceMapperService/addressProcessing.test.js
@@ -12,7 +12,7 @@ describe('Address Processing', () => {
         CustomerName: { content: 'Test Customer' }
       };
       
-      const result1 = mapper.extractCustomerData(fields);
+      const result1 = mapper.EntityExtractor.extractCustomerData(fields);
       expect(result1.address).toBe('123 Main St Austin, TX 78701');
       expect(result1.name).toBe('Test Customer');
       
@@ -24,7 +24,7 @@ describe('Address Processing', () => {
         CustomerName: { content: 'Test Customer 2' }
       };
       
-      const result2 = mapper.extractCustomerData(fields2);
+      const result2 = mapper.EntityExtractor.extractCustomerData(fields2);
       expect(result2.address).toBe('456 Oak Ave Seattle WA 98101');
       
       // International format
@@ -35,7 +35,7 @@ describe('Address Processing', () => {
         CustomerName: { content: 'Test Customer 3' }
       };
       
-      const result3 = mapper.extractCustomerData(fields3);
+      const result3 = mapper.EntityExtractor.extractCustomerData(fields3);
       expect(result3.address).toBe('789 High Street London, England EC1A 1BB');
     });
     
@@ -51,7 +51,7 @@ describe('Address Processing', () => {
         BillingAddressRecipient: { content: 'John Doe' }
       };
       
-      const result = mapper.extractCustomerData(fields);
+      const result = mapper.EntityExtractor.extractCustomerData(fields);
       expect(result.address).toBe('123 Main Avenue Boston, MA 02108');
     });
     
@@ -63,7 +63,7 @@ describe('Address Processing', () => {
         // No address provided
       };
       
-      const result = mapper.extractCustomerData(fields);
+      const result = mapper.EntityExtractor.extractCustomerData(fields);
       expect(result.address).toBeNull();
       expect(result.name).toBe('Test Customer');
     });
@@ -79,7 +79,7 @@ describe('Address Processing', () => {
         VendorTaxId: { content: '123-45-6789' }
       };
       
-      const result = mapper.extractVendorData(fields);
+      const result = mapper.EntityExtractor.extractVendorData(fields);
       expect(result.address).toBe('123 Business St Chicago, IL 60601');
       expect(result.name).toBe('Test Vendor');
       expect(result.tax_id).toBe('123-45-6789');
